# backend/backend/docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Sari Tebu API",
    "description": "Dokumentasi REST API untuk Sari Tebu, aplikasi POS (Point of Sale) berbasis web. Mencakup pendaftaran akun dengan verifikasi email, manajemen produk, keranjang belanja, sampai proses checkout.\n\nResponse selalu dalam format JSON dengan field status: \"success\" kalau request berhasil, \"fail\" kalau ada kesalahan dari sisi client (misalnya validasi gagal atau resource-nya nggak ada), dan \"error\" kalau ada masalah di server yang nggak terduga.\n\nAutentikasi di sini pakai session token yang disimpan di HTTP-only cookie, jadi bukan Bearer token/JWT seperti yang biasa dipakai di banyak API lain. Karena itu kebanyakan endpoint nggak perlu header Authorization.\n\nUntuk ID resource, formatnya pakai prefix sesuai jenisnya, contoh user-xxxxxxxxxxxxxxxxxxxxx atau product-xxxxxxxxxxxxxxxxxxxxx, di mana bagian setelah prefix itu di-generate pakai nanoid (21 karakter).",
    "version": "1.0.0",
    "contact": {
      "name": "Universitas Mikroskil IF-A Sore Kelompok A",
      "url": "https://github.com/Mikroskil-IFA-Sore-A/Sari-Tebu"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
        "url": "http://localhost:8080/api",
        "description": "Local Development server"
    },
    {
        "url": "https://saritebu.aelberthcheong.dev/api",
        "description": "Production server"
    }
  ],
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["message"]
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^product-[A-Za-z0-9_-]{21}$",
            "description": "Unique product identifier"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "price": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "description": "Product price"
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "description": "Available stock quantity"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Product creation timestamp"
          }
        },
        "required": ["id", "name", "price", "stock"]
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "pattern": "^product-[A-Za-z0-9_-]{21}$",
            "description": "Product identifier"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantity in cart"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": ["product_id", "quantity"]
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^cart-[A-Za-z0-9_-]{21}$",
            "description": "Unique cart identifier"
          },
          "user_id": {
            "type": "string",
            "pattern": "^user-[A-Za-z0-9_-]{21}$",
            "description": "User who owns the cart"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "user_id", "items"]
      },
      "TransactionItem": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "pattern": "^product-[A-Za-z0-9_-]{21}$",
            "description": "Product identifier"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "price_at_time": {
            "type": "number",
            "format": "float",
            "description": "Product price at the moment of checkout"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": ["product_id", "quantity", "price_at_time"]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^transaction-[A-Za-z0-9_-]{21}$",
            "description": "Unique transaction identifier"
          },
          "user_id": {
            "type": "string",
            "pattern": "^user-[A-Za-z0-9_-]{21}$",
            "description": "User who made the transaction"
          },
          "total": {
            "type": "number",
            "format": "float",
            "description": "Total price of the transaction"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionItem"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "user_id", "total", "items"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_session_token"
      }
    }
  },
  "paths": {
...
    "/carts": {
      "get": {
        "tags": ["Cart"],
        "summary": "List carts",
        "description": "Retrieve all carts for the authenticated user.",
        "operationId": "listCarts",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of carts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cart"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Cart"],
        "summary": "Create cart",
        "description": "Create a new shopping cart for the authenticated user.",
        "operationId": "createCart",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Cart created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/carts/{cartId}": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get cart",
        "description": "Retrieve details of a specific cart.",
        "operationId": "getCart",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "description": "Cart ID",
            "schema": {
              "type": "string",
              "pattern": "^cart-[A-Za-z0-9_-]{21}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Delete cart",
        "description": "Delete a cart and all its items.",
        "operationId": "deleteCart",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "description": "Cart ID",
            "schema": {
              "type": "string",
              "pattern": "^cart-[A-Za-z0-9_-]{21}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cart deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "List transactions",
        "description": "Retrieve all transactions for the authenticated user.",
        "operationId": "listTransactions",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Transactions"],
        "summary": "Checkout",
        "description": "Process a checkout transaction from an existing cart. This empties and deletes the cart, decrements product stock, and creates a transaction record.",
        "operationId": "checkout",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartId": {
                    "type": "string",
                    "pattern": "^cart-[A-Za-z0-9_-]{21}$",
                    "description": "Cart to check out"
                  }
                },
                "required": ["cartId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Cart not found or does not belong to the user"
          }
        }
      }
    },
    "/transactions/{transactionId}": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get transaction",
        "description": "Retrieve details of a specific transaction.",
        "operationId": "getTransaction",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "description": "Transaction ID",
            "schema": {
              "type": "string",
              "pattern": "^transaction-[A-Za-z0-9_-]{21}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  }
}